requirements.txt
youtube-transcript-api>=1.0.0
google-genai>=1.0.0
flask>=3.0.0
